// src/app/dashboard/types/TypePartyStatement.ts
export interface TypePartyStatement {
    RowNo: number;
    TransSno: number;
    Trans_Date: Date;
    SeriesSno: number;
    Particulars: string;
    CrAmount: number;
    DrAmount: number;
    CrWeight: number;
    DrWeight: number;
    Remarks: string;
    Balance: number;
    BalanceWt: number;
}